Maaaaaaay have overinvested on this one...
I have to confess, the original link was missing because I skipped this lesson as being too easy and just wrote NA (and hey, they're optional, right?).  I did that so it would appear as complete on my end, but didn't think about the fact that someone would have to grade it.  Well, you wanted a list?  You got one!